inventory-traceroute-agent/requirements.txt
# Python dependencies for your agent
# Add any packages your agent needs here

# Example dependencies:
# requests>=2.28.0
# pandas>=1.5.0
# numpy>=1.21.0


# Core Agent Framework & Metrics API
ncp-sdk>=0.1.0

# SSH Connectivity for Traceroute
paramiko>=2.11.0

# Visualization & Graphing
networkx>=3.1
plotly>=5.15.0
pandas>=2.0.0
pyvis
